pyproject.toml
[project]
name = "seeing-sound"
version = "0.1.0"
description = ""
authors = [
    {name = "Mike Robinson",email = "michael_robinson@live.ca"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyaudio (>=0.2.14,<0.3.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pyqt6 (>=6.9.0,<7.0.0)",
    "lifxlan @ git+ssh://git@github.com:mclarkk/lifxlan.git"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
run = "seeing_sound.main:main"
run_gui = "seeing_sound.gui:main"